requirements.txt
PyYAML==3.10
git+https://github.com/rgrp/pywordpress.git